Add files via upload
implemented GoogleMaps API feature complete with features such as geolocation, distance between coordinates, etc. Next step in the process is to overlay widgets on top of the map, such as coordinates and icons for favorites, friends, enemies (social networking portion of the map), and call features of the API to augment the features of the application.